% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-label.R
\name{cal_label}
\alias{cal_label}
\title{Construct a label object for calendar heatmap parameters}
\usage{
cal_label(position = c("bottom", "top", "left", "right"),
  align = c("center", "left", "right"), rotate = c("none", "left",
  "right"), offset = cal_offset(0, 0), width = 100, height = NULL)
}
\arguments{
\item{position}{position of the label, relative to the \code{domain} one of
\code{top}, \code{left}, \code{bottom} or \code{right} (or abbreviated versions). Default
is \code{bottom}.}

\item{align}{horizontal alignment of the \code{domain}. One of \code{left}, \code{right} or
\code{center}. Default is \code{center}.}

\item{rotate}{rotation of vertical label. One of \code{none}, \code{left} or \code{right}.
Defautl is \code{none}.}

\item{offset}{additional control over label positioning. i.e. if the default
does not fit your need (likley when rotation is involved) or when
using a larger font size. Use \code{\link[=cal_offset]{cal_offset()}} to construct. Defualt
is \code{cal_offset(0,0)}.}

\item{width}{\emph{only} used when \code{roate} is not \code{none}! This defines the width
of the label (in pixels).}

\item{height}{height of the \code{domain} labels in pixels. If \code{NULL} the label
is automagically set to 2x the height of \code{sub_domain} cells.}
}
\description{
Construct a label object for calendar heatmap parameters
}
