% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendary.R
\name{cal_legend}
\alias{cal_legend}
\title{Modify the calendar heatmap legend}
\usage{
cal_legend(cwdgt, show = TRUE, breaks = 5, colors = NULL,
  cell_size = 10, padding = 2, margin = cal_margin(10),
  vjust = "bottom", hjust = "top", orientation = "horizontal")
}
\arguments{
\item{cwdgt}{a calendar widget object}

\item{show}{if \code{TRUE} show the legend. Default: \code{TRUE}}

\item{breaks}{number of breaks for the legend values; Default is 5}

\item{colors}{colors to compute the legend from. Use \code{\link[=cal_colors]{cal_colors()}} to construct.
Default is \code{NULL} so you get that horrible greeen legend.}

\item{cell_size}{the size of the legend cells in pixels. Default is 10 pixels.}

\item{padding}{the padding between each legen cell in pixels. Default is 2 pixels.}

\item{margin}{the margin around the legend. Use \code{\link[=cal_margin]{cal_margin()}} to set this.
Default is \code{cal_margin(10)}.}

\item{vjust}{the vertical positon of the legend. One of \code{bottom}, \code{center} or
\code{top}. Default is \code{bottom}.}

\item{hjust}{the horizontal position of the legend. One of \code{left}, \code{center} or
\code{right}. Default is \code{left}.}

\item{`orientation`}{the orientation of the legend. One of \code{horizontal} or
\code{vertical}. Default is \code{horizontal}.}
}
\description{
Modify the calendar heatmap legend
}
