% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-margin.R
\name{cal_margin}
\alias{cal_margin}
\title{Construct a margin object for calendar heatmap parameters}
\usage{
cal_margin(t = 0, r = 0, b = 0, l = 0)
}
\arguments{
\item{t, r, b, l}{top, bottom, left, right margins (in pixels)}
}
\description{
Construct a margin object for calendar heatmap parameters
}
