% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-names.R
\name{cal_names}
\alias{cal_names}
\title{Construct a tooltip names object for the calender widget}
\usage{
cal_names(singular = "item", plural = sprintf("\%ss", singular))
}
\arguments{
\item{singular, plural}{singular and plural forms for the "thing(s)".
Defaults to naively sticking an \code{s} on the end of \code{singular} and
defaults \code{singular} to \code{item}.}
}
\description{
Construct a tooltip names object for the calender widget
}
