% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-offset.R
\name{cal_offset}
\alias{cal_offset}
\title{Construct an offset object for calendar labels}
\usage{
cal_offset(x = 0, y = 0)
}
\arguments{
\item{x, y}{x and y offsets}
}
\description{
Construct an offset object for calendar labels
}
