% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal-style.R
\name{cal_style}
\alias{cal_style}
\title{Add CSS styling rules to a calendar heatmap}
\usage{
cal_style(cwdgt, subdomain_text, hover_cursor, graph_label, graph_rect,
  highlight, highlight_now, rect_highlight, rect_highlight_now,
  text_highlight, text_highlight_now)
}
\arguments{
\item{cwdgt}{a calendar widget object}

\item{subdomain_text}{CSS style for \code{sub_domain} text. Note you can't change
horizontal or vertical justification.}

\item{hover_cursor}{CSS for all clickable cells}

\item{graph_label}{CSS for the \code{domain} label}

\item{graph_rect}{CSS for the \code{sub_domain} cells and used when the cell is not
associated to a value.}

\item{highlight, highlight_now}{CSS for higlighted cells}

\item{rect_highlight, rect_highlight_now, text_highlight, text_highlight_now}{used
instead of \code{highlight}/\code{highlight_now} to set CSS on the cell rect and the text
inside respectively.}
}
\description{
Add CSS styling rules to a calendar heatmap
}
