% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-cells.R
\name{label_cells}
\alias{label_cells}
\title{Display labels on calendar heatmap cells}
\usage{
label_cells(cwdgt, format_string = NULL, js_body = NULL)
}
\arguments{
\item{cwdgt}{a calendar widget object}

\item{format_string}{a valid javascript date format string \emph{OR}}

\item{js_body}{bare javascript function body code that will be wrapped in
\code{function(date ,value) {}} for you. You must return a value
that can be converted to a string.}
}
\description{
Use one of \code{format_string} (for simple data format strings) or
javascript function body code to use other values. Not both.
}
\references{
\href{https://cal-heatmap.com/#subDomainTextFormat}{subdomainTextFormat}
}
