% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercaliheatmapwidget-package.R
\docType{package}
\name{supercaliheatmapwidget}
\alias{supercaliheatmapwidget}
\alias{supercaliheatmapwidget-package}
\title{Supercalifragilistic HTML Calendar Heatmaps}
\description{
The veritable Mary Poppins of interactive R calendar heatmaps.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/supercaliheatmapwidget}
\item BugReports: \url{https://gitlab.com/hrbrmstr/supercaliheatmapwidget/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
