% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supercaliheatmapwidget-package.R,
%   R/supercaliheatmapwidget.R
\docType{package}
\name{supercaliheatmapwidget}
\alias{supercaliheatmapwidget}
\alias{supercaliheatmapwidget-package}
\title{Supercalifragilistic HTML Calendar Heatmaps}
\usage{
supercaliheatmapwidget(xdf, datetime_col, value_col, tooltip = TRUE,
  tooltip_item_name = cal_names(), domain = c("month", "day", "hour",
  "week", "year"), sub_domain = c("day", "min", "x_min", "hour",
  "x_hour", "x_day", "week", "x_week", "month", "x_month"), range = 12,
  start = NULL, cell_size = 20, cell_padding = 4, cell_radius = 0,
  domain_gutter = 6, domain_margin = cal_margin(),
  orientation = c("horizontal", "vertical"), label = cal_label(),
  col_limit = NULL, row_limit = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{xdf}{data frame}

\item{datetime_col}{bare or quoted column name where \code{Date} or \code{POSIXct}
(or equivalent character strings) are stored.}

\item{value_col}{bare or quoted column name where integer or numeric values
are stored.}

\item{tooltip}{display a tooltip on hover? Default is \code{TRUE}.}

\item{tooltip_item_name}{the "thing" being counted in the heatmap cells.
Use \code{\link[=cal_names]{cal_names()}} to set the singular and plural form of the "thing".
Defaults to \code{cal_names('item', 'items')}}

\item{domain}{the type of "domain" for the calendar. One of \code{hour}, \code{day},
\code{week}, \code{month} or \code{year}; Defaults to \code{month}. See \href{https://cal-heatmap.com/#domain}{the docs}.}

\item{sub_domain}{the type of "subdomain" for the "domain" of the calendar. This
has to make sens efor the choise of "domain". One of \code{min}, \code{x_min}, \code{hour},
\code{x_hour}, \code{day}, \code{x_day}, \code{week}, \code{x_week}, \code{month} or \code{x_month}. Defaults to \code{day}
since \code{domain} defaults to \code{month}. See \href{https://cal-heatmap.com/#subdomain}{the docs}.}

\item{range}{the number of \code{domain} to display. The default is \code{12} (e.g. 12 months).}

\item{start}{the start date of the calendar. Can be a \code{Date} or \code{POSIXct} (or
equivalent character value but must match the type/format of \code{datetime_col}).
See \href{https://cal-heatmap.com/#start}{the docs}.}

\item{cell_size}{the size of each \code{sub_domain} cell in pixels. Default is \code{20}.}

\item{cell_padding}{the space between each \code{sub_domain} cell in pixels. Default is \code{4}.}

\item{cell_radius}{the \code{sub_domain} cell's border radius in pixels. This enables
rounded corners. Default is \code{0} or a "box".}

\item{domain_gutter}{the space betwen each \code{domain} in pixels. Defaults to \code{6}.}

\item{domain_margin}{the top, bottom, left, right margins around each \code{domain} in
pixels. Use \code{\link[=cal_margin]{cal_margin()}} to construct this.}

\item{orientation}{default is "\code{horizontal}". Can also be "\code{vertical}".}

\item{label}{position and alignment of the \code{domain} label. USe \code{\link[=cal_label]{cal_label()}} to construct.}

\item{col_limit}{control the number of columns to split the \code{domain} into. Each domain
is split into an arbitrary number of columns (or rows depending on the reading
direction). You can overwrite that number with this parameter and force all
dates on the same line, or split them into more columns. Default is \code{NULL}.}

\item{row_limit}{control the number of rows to split the \code{domain} into. Default is \code{NULL}.`}

\item{width, height, elementId}{the usual \code{htmlwidget} suspsects.}
}
\description{
The veritable Mary Poppins of interactive R calendar heatmaps.

The veritable Mary Poppins of interactive R calendar heatmaps.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/supercaliheatmapwidget}
\item BugReports: \url{https://gitlab.com/hrbrmstr/supercaliheatmapwidget/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
